requirements.txt
matplotlib
numpy
scikit-image
tensorflow-gpu
